+++ src/main/java/com/resource/api/dao/IProductoDao.java
@@ -6,8 +6,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.resource.api.entity.Producto;
 
-public interface IProductoDao extends CrudRepository<Producto,Integer>{
+public interface IProductoDao extends CrudRepository<Producto, Long>{
 	
-	public List<Producto>findByProneg(Integer negid);
+	public List<Producto>findByProneg(Long negid);
 	
 }
