+++ src/main/java/com/resource/api/dao/IUsuarioDao.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.resource.api.entity.Usuario;
 
-public interface IUsuarioDao extends CrudRepository<Usuario,Integer>{
+public interface IUsuarioDao extends CrudRepository<Usuario, Long>{
 	
 	public Usuario findByUsuemail(String usuemail);
-	public Optional<Usuario>findById(Integer usuid);
+	public Optional<Usuario>findById(Long usuid);
 	
 }
